changed all int types to doubles, build parser for response from url

+++ tomatoAI/Plant.java
@@ -3,7 +3,7 @@ public class Plant {
     String name = "";
     String plantType = "";
     String weather = "";
-    int plantAge = 0;
+    double plantAge = 0;
     double plantSize = 0;
     double waterConsumption = 0;
     double loamAmount = 0;
@@ -11,7 +11,7 @@ public class Plant {
     double sandAmount = 0;
     double hoseFlow = 0;
 
-    Plant(String name, String plantType, String weather, int plantAge, double plantSize, double waterConsumption, double loamAmount, double clayAmount, double sandAmount, double hoseFlow) {
+    Plant(String name, String plantType, String weather, double plantAge, double plantSize, double waterConsumption, double loamAmount, double clayAmount, double sandAmount, double hoseFlow) {
         this.name = name;
         this.plantType = plantType;
         this.weather = weather;
@@ -40,11 +40,11 @@ public class Plant {
         return weather;
     }
 
-    public int getPlantAge() {
+    public double getPlantAge() {
         return plantAge;
     }
 
-    public void setPlantSize(int newSize) {
+    public void setPlantSize(double newSize) {
         plantSize = newSize;
     }
 
@@ -76,7 +76,7 @@ public class Plant {
         return hoseFlow;
     }
 
-    public void setNewSoilComp(int newLoam, int newClay, int newSand) {
+    public void setNewSoilComp(double newLoam, double newClay, double newSand) {
         loamAmount = newLoam;
         clayAmount = newClay;
         sandAmount = newSand;
